Fix GUI Scaling
I started looking into the GUI scaling problem I reported: 
https://forum.kerbalspaceprogram.com/index.php?/topic/163315-151-trp-hire-formerly-ksi-hiring/&do=findComment&comment=3524527

One of the first things I did was comment out the scaling stuff that was already in this routine and it solved the problem. So this PR is stripping the scaling "fixes" that were already in place. Tested locally at gui scale = 80%, 100% and 140%